replacing decimal2p with custom version removing unnecessary console prints improving documentation restructuring package

+++ controller/monthly_payments.go
@@ -44,7 +44,7 @@ func MonthlyPayments(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	log.Printf("Calculating monthly payments for %.2f for property with initial market value %.2f to be paid in %d months with automatic updates %v and Extra Repayments %v", rb.InitialPrincipal, rb.MarketValue.AsFloat64(), rb.Months, rb.AutomaticInterestUpdate, rb.Repayments)
+	log.Printf("Calculating monthly payments for %.2f for property with initial market value %.2f to be paid in %d months with automatic updates %v and Extra Repayments %v", rb.InitialPrincipal, rb.MarketValue.AsFloat(), rb.Months, rb.AutomaticInterestUpdate, rb.Repayments)
 
 	mps := monthlypayments.CalculateLinearMonthlyPayments(rb)
